[#34496] fix linter complaints

+++ bibfor/include/asterfort/jeveuo.h
@@ -1,5 +1,5 @@
 ! --------------------------------------------------------------------
-! Copyright (C) 1991 - 2023 - EDF R&D - www.code-aster.org
+! Copyright (C) 1991 - 2025 - EDF R&D - www.code-aster.org
 ! This file is part of code_aster.
 !
 ! code_aster is free software: you can redistribute it and/or modify
@@ -24,10 +24,10 @@
 #include "asterf_types.h"
     character(len=*), intent(in) :: nomlu
     character(len=*), intent(in) :: cel
-    integer, optional :: jadr
+    integer(kind=8), optional :: jadr
 
     aster_logical,     pointer, optional :: vl(:)
-    integer,           pointer, optional :: vi(:)
+    integer(kind=8),           pointer, optional :: vi(:)
     integer(kind=4),   pointer, optional :: vi4(:)
     real(kind=8),      pointer, optional :: vr(:)
     complex(kind=8),   pointer, optional :: vc(:)
